[RAPTOR-8462] Unify the flows for handling models and deployments into a single one
