--- .github/actions/custom-inference-model-deployment/action.yml
@@ -1,64 +1,0 @@
-# action.yml
-name: 'Custom Inference Model Deployment'
-description: 'Manage custom inference model deployments in DataRobot'
-inputs:
-  api-token: # id of input
-    description: 'DataRobot authentication API token'
-    required: true
-  webserver: # id of input
-    description: 'DataRobot frontend web server'
-    required: true
-  skip-cert-verification: # id of input
-    description: 'Whether a request to an HTTPS URL will be made without a certificate verification'
-    required: false
-    default: 'false'
-  release-branch:  # id of input
-    description: |
-      The release branch for which pull request and push events will trigger the
-      action.
-    required: true
-  allow-deployment-deletion:
-    description: |
-      Whether to detect local deleted deployment definitions and consequently delete them
-      in DataRobot.
-    required: false
-    default: 'false'
-outputs:
-  total-affected-deployments: # id of output
-    description: 'How many deployments were affected.'
-    value: ${{ steps.custom-inference-model-deployment.outputs.total-affected-deployments }}
-  total-created-deployments: # id of output
-    description: 'How many new deployments were created.'
-    value: ${{ steps.custom-inference-model-deployment.outputs.total-created-deployments }}
-  total-deleted-deployments: # id of output
-    description: 'How many deployments were deleted.'
-    value: ${{ steps.custom-inference-model-deployment.outputs.total-deleted-deployments }}
-  message: # id of output
-    description: 'The output message from the GitHub action.'
-    value: ${{ steps.custom-inference-model-deployment.outputs.message }}
-runs:
-  using: 'composite'
-  steps:
-    - run: pip install -r ${GITHUB_ACTION_PATH}/../../../src/requirements.txt
-      shell: bash
-    - id: custom-inference-model-deployment
-      run: |
-        if ${{ inputs.skip-cert-verification }} == true; then
-        verify_cert_arg="--skip-cert-verification"
-        else
-        verify_cert_arg=""
-        fi
-        if ${{ inputs.allow-deployment-deletion }} == true; then
-        allow_deployment_deletion_arg="--allow-deployment-deletion"
-        else
-        allow_deployment_deletion_arg=""
-        fi
-        python ${GITHUB_ACTION_PATH}/../../../src/main.py \
-          --deploy \
-          --api-token ${{ inputs.api-token }} \
-          --webserver ${{ inputs.webserver }} \
-          ${verify_cert_arg} \
-          --branch ${{ inputs.release-branch }} \
-          --root-dir "${{ github.workspace }}" \
-          ${allow_deployment_deletion_arg}
-      shell: bash
