--- .github/actions/custom-inference-model/action.yml
@@ -1,66 +1,0 @@
-# action.yml
-name: 'Custom Inference Model'
-description: 'Manage custom inference models in DataRobot'
-inputs:
-  api-token: # id of input
-    description: 'DataRobot authentication API token.'
-    required: true
-  webserver: # id of input
-    description: 'DataRobot frontend web server.'
-    required: true
-  skip-cert-verification: # id of input
-    description: |
-      Whether a request to an HTTPS URL will be made without a certificate verification.
-    required: false
-    default: 'false'
-  main-branch:  # id of input
-    description: 'The main branch for which pull request and push events will trigger the action.'
-    required: false
-    default: 'master'
-  allow-model-deletion:
-    description:  |
-      Whether to detected local deleted model definitions and consequently delete them
-      in DataRobot.
-    required: false
-    default: 'false'
-outputs:
-  total-affected-models: # id of output
-    description: 'The total number of models that were affected.'
-    value: ${{ steps.custom-inference-model.outputs.total-affected-models }}
-  total-created-models: # id of output
-    description: 'The total number of new models that were created.'
-    value: ${{ steps.custom-inference-model.outputs.total-created-models }}
-  total-deleted-models: # id of output
-    description: 'The total number of models that were deleted.'
-    value: ${{ steps.custom-inference-model.outputs.total-deleted-models }}
-  total-created-model-versions: # id of output
-    description: 'The total number of new model versions that were created.'
-    value: ${{ steps.custom-inference-model.outputs.total-created-model-versions }}
-  message: # id of output
-    description: 'The output message from the GitHub action.'
-    value: ${{ steps.custom-inference-model.outputs.message }}
-runs:
-  using: 'composite'
-  steps:
-    - run: pip install -r ${GITHUB_ACTION_PATH}/../../../src/requirements.txt
-      shell: bash
-    - id: custom-inference-model
-      run: |
-        if ${{ inputs.skip-cert-verification }} == true; then
-          verify_cert_arg="--skip-cert-verification"
-        else
-          verify_cert_arg=""
-        fi
-        if ${{ inputs.allow-model-deletion }} == true; then
-          allow_model_deletion_arg="--allow-model-deletion"
-        else
-          allow_model_deletion_arg=""
-        fi
-        python ${GITHUB_ACTION_PATH}/../../../src/main.py \
-          --api-token ${{ inputs.api-token }} \
-          --webserver ${{ inputs.webserver }} \
-          ${verify_cert_arg} \
-          --branch ${{ inputs.main-branch }} \
-          --root-dir "${{ github.workspace }}" \
-          ${allow_model_deletion_arg}
-      shell: bash
