+++ .github/workflows/workflow.yml
@@ -83,8 +83,6 @@ jobs:
     # Run this job on any action of a PR, but skip the job upon merge to master
     if: github.event.pull_request.merged != true
 
-    needs: validate-code-style
-
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
 
@@ -100,7 +98,7 @@ jobs:
         run: make test
 
   run-functional-test:
-    needs: run-unit-tests
+    needs: [run-unit-tests, validate-code-style]
 
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
@@ -116,9 +114,8 @@ jobs:
       - name: Run a functional test
         run: |
           make test-functional \
-            FUNCTIONAL_TESTS='tests/functional/test_deployment_github_actions.py::TestDeploymentGitHubActions::test_e2e_deployment_create[push]'
+            FUNCTIONAL_TESTS='tests/functional/test_deployment_github_actions.py::TestDeploymentGitHubActions::test_e2e_deployment_create'
 
-  # This workflow contains a single job called "build"
   datarobot-custom-inference-model:
     # Skip this job! It is an intermediate step towards putting it in a separate repository.
     # Run this job on any action of a PR, but skip the job upon merge to master
@@ -153,69 +150,26 @@ jobs:
         run: git log --name-only
 
       - name: DataRobot Custom Inference Model
-        id: custom-inference-model
+        id: custom-inference-model-and-deployment
         if: ${{ env.INSPECT_CONTEXT == 'false' }}
-        uses: ./.github/actions/custom-inference-model
+        uses: ./
         with:
           api-token: $DATAROBOT_API_TOKEN
           webserver: $DATAROBOT_WEBSERVER
-          main-branch: master
-          skip-cert-verification: true
-
-      - name: DataRobot Custom Inference Model Results
-        run: |
-          echo "Total affected models: ${{ steps.custom-inference-model.outputs.total-affected-models }}"
-          echo "Total created models: ${{ steps.custom-inference-model.outputs.total-created-models }}"
-          echo "Total deleted models: ${{ steps.custom-inference-model.outputs.total-deleted-models }}"
-          echo "Total created model versions: ${{ steps.custom-inference-model.outputs.total-created-model-versions }}"
-          echo "Message: ${{ steps.custom-inference-model.outputs.message }}"
-
-  datarobot-custom-inference-model-deployment:
-    needs: [validate-code-style, run-unit-tests, datarobot-custom-inference-model]
-
-    # Skip this job! It is an intermediate step towards putting it in a separate repository.
-    if: ${{ false }}
-
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - name: GitHub Environment Variables
-        run: |
-          echo "GITHUB_EVENT_NAME: ${GITHUB_EVENT_NAME}"
-          echo "GITHUB_SHA: ${GITHUB_SHA}"
-          echo "GITHUB_HEAD_REF: ${GITHUB_HEAD_REF}"
-          echo "GITHUB_BASE_REF: ${GITHUB_BASE_REF}"
-          echo "GITHUB_REF: ${GITHUB_REF}"
-          echo "GITHUB_REF_NAME: ${GITHUB_REF_NAME}"
-          echo "GITHUB_REF_PROTECTED: ${GITHUB_REF_PROTECTED}"
-          echo "GITHUB_REF_TYPE: ${GITHUB_REF_TYPE}"
-          echo "GITHUB_RUN_ID: ${GITHUB_RUN_ID}"
-          echo "GITHUB_RUN_NUMBER: ${GITHUB_RUN_NUMBER}"
-
-      - name: Git log
-        run: git log --name-only
-
-      - name: DataRobot Custom Inference Model Deployment
-        id: custom-inference-model-deployment
-        if: ${{ env.INSPECT_CONTEXT == 'false' }}
-        uses: ./.github/actions/custom-inference-model-deployment
-        with:
-          api-token: $DATAROBOT_API_TOKEN
-          webserver: $DATAROBOT_WEBSERVER
-          release-branch: master
-          skip-cert-verification: true
+          branch: master
+          allow-model-deletion: true
           allow-deployment-deletion: true
+          skip-cert-verification: true
 
-      - name: DataRobot Custom Inference Model Deployment Results
+      - name: DataRobot Custom Inference Model Action Results
         run: |
-          echo "Total affected deployments: ${{ steps.custom-inference-model-deployment.outputs.total-affected-deployments }}"
-          echo "Total created deployments: ${{ steps.custom-inference-model-deployment.outputs.total-created-deployments }}"
-          echo "Total deleted deployments: ${{ steps.custom-inference-model-deployment.outputs.total-deleted-deployments }}"
-          echo "Message: ${{ steps.custom-inference-model-deployment.outputs.message }}"
+          echo "Total affected models: ${{ steps.custom-inference-model-and-deployment.outputs.total-affected-models }}"
+          echo "Total created models: ${{ steps.custom-inference-model-and-deployment.outputs.total-created-models }}"
+          echo "Total deleted models: ${{ steps.custom-inference-model-and-deployment.outputs.total-deleted-models }}"
+          echo "Total created model versions: ${{ steps.custom-inference-model-and-deployment.outputs.total-created-model-versions }}"
+
+          echo "Total affected deployments: ${{ steps.custom-inference-model-and-deployment.outputs.total-affected-deployments }}"
+          echo "Total created deployments: ${{ steps.custom-inference-model-and-deployment.outputs.total-created-deployments }}"
+          echo "Total deleted deployments: ${{ steps.custom-inference-model-and-deployment.outputs.total-deleted-deployments }}"
+
+          echo "Message: ${{ steps.custom-inference-model-and-deployment.outputs.message }}"
